feat: move topic recap endpoint
